# Minor Optimizations in Radix Cache and Schedule Batch