<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

This PR introduces performance optimizations to two key components of the SGLang serving stack: `radix_cache.py` and `schedule_batch.py`. The primary goal is to reduce latency and improve throughput by eliminating inefficient data structures and unnecessary computations during request processing and KV cache eviction.

## Modifications

### Optimizations in radix_cache.py
**1. Reduced latency via optimized eviction logic**
Before: Average latency over 12 requests: 0.577s
After: Average latency over 12 requests: 0.495s

**Evict Method Improvements (~60% speedup, from 31ms → 12ms):**

- Avoid rebuilding the heap with `heapq.heapify()` on every eviction
    - Previously, the eviction heap was rebuilt from scratch each time (O(n) cost).
    - ✅ Now: Maintain a persistent min-heap (self.evict_heap) and incrementally update it during insertions, lock/unlock updates, and deletions.

- Avoid full-tree traversal in `_collect_leaves()`
    - Tree traversal for collecting evictable leaves was costly and unnecessary on each eviction.
    - ✅ Now: The persistent heap contains all evictable leaves, eliminating the need for traversal.

- Avoid repeated heap rebalancing
    - Each heapq.heappop() call triggered rebalancing.
    - ✅ Now: evict() loop pops only unlocked and valid nodes using a guard, minimizing operations.

**Implementation Details:**
- Added self.evict_heap: List[Tuple[int, TreeNode]] to RadixCache.
- Heap is updated incrementally:
    - insert(): Add leaf if evictable.
    - inc_lock_ref(): Remove node from heap if it becomes locked.
    - dec_lock_ref(): Re-add node if it becomes unlockable.
    - _delete_leaf(): No-op for heap; eviction logic handles parent-child promotion.

**2. Faster leaf deletion via key tracking**
Previously, _delete_leaf() required a linear search in the parent’s children dict to find the key corresponding to the child node.
✅ Now: Each TreeNode stores its child_key upon insertion, eliminating the need to search during deletion.

### Optimizations in schedule_batch.py
Before: Average latency over 12 requests: 0.495s
After: Average latency over 12 requests: 0.475s

**Improvements:**
1. Flattening input IDs efficiently
❌ Before: `sum(input_ids, [])` (O(n²) operation due to repeated list copying).
✅ Now: `list(chain.from_iterable(input_ids))` for linear-time flattening.

2. Avoid redundant tensor construction
❌ Before: `torch.tensor(seq_lens, ...)` may create unnecessary copies.
✅ Now: `torch.as_tensor(seq_lens, dtype=torch.int64, device=self.device)` avoids extra copy if input is already a tensor-like object.

These optimizations contribute to overall reduced latency and improved batch throughput, especially under high request volume and large prompt size settings.

## Checklist

- [X] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [X] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.

##Profiling Screenshots:
Radix Cache:
BEFORE:
![BEFORE - Radix Cache Profiler Screenshot](https://github.com/user-attachments/assets/09e8b16c-2810-418f-8220-5d2c2a6f7881)
AFTER:
![AFTER - Radix Cache Profiler Screenshot](https://github.com/user-attachments/assets/6b526a5f-2e64-4261-8102-be1adce3ab42)

Schedule Batch Logic:
BEFORE:
![BEFORE - Schedule Batch Profiler Screenshot](https://github.com/user-attachments/assets/88c855fd-2dc9-43c7-a64e-779004c2879c)
AFTER:
![AFTER - Schedule Batch Profiler Screenshot](https://github.com/user-attachments/assets/c5f92a55-3692-4710-8583-245a60650684)