@merrymercy 2025-08-03
We will close this one. Feel free to reopen if you think it is still necessary.

- Please try this c++ implementation. It should be much faster https://github.com/sgl-project/sglang/pull/7369
- @hanming-lu tested this on our benchmark but did not found noticeable perf difference between this and main. Could you tell us the exact instructions on how to reproduce?